bug(UIKIT-982): Рефакторинг для типов, без переноса в types
