+++ packages/components/src/Badge/Badge.tsx
@@ -1,5 +1,23 @@
-import { BadgeProps } from './types';
+import { BadgeProps as MuiBadgeProps } from '@mui/material/Badge/Badge';
+
+import { WithoutEmotionSpecific } from '../types';
+
+import { BadgeColor } from './types';
 import { StyledBadge } from './styles';
+
+export type BadgeProps = Omit<
+  WithoutEmotionSpecific<MuiBadgeProps>,
+  'color'
+> & {
+  /**
+   * Цвет фона
+   */
+  color: BadgeColor;
+  /**
+   * Наличие белой рамки
+   */
+  withBorder?: boolean;
+};
 
 export const Badge = ({
   children,
