+++ packages/components/src/Badge/types.ts
@@ -1,7 +1,3 @@
-import { BadgeProps as MuiBadgeProps } from '@mui/material/Badge/Badge';
-
-import { WithoutEmotionSpecific } from '../types';
-
 export type BadgeColor =
   | 'primary'
   | 'grey'
@@ -10,17 +6,3 @@ export type BadgeColor =
   | 'errorLight'
   | 'error'
   | 'success';
-
-export type BadgeProps = Omit<
-  WithoutEmotionSpecific<MuiBadgeProps>,
-  'color'
-> & {
-  /**
-   * Цвет фона
-   */
-  color: BadgeColor;
-  /**
-   * Наличие белой рамки
-   */
-  withBorder?: boolean;
-};
