+++ packages/components/src/Checkbox/index.ts
@@ -1,3 +1,1 @@
 export * from './Checkbox';
-
-export * from './types';
