--- packages/components/src/Checkbox/types.ts
@@ -1,13 +1,0 @@
-import { CheckboxProps as MuiCheckboxProps } from '@mui/material/Checkbox/Checkbox';
-
-import { WithoutEmotionSpecific } from '../types';
-
-export type CheckboxProps = Omit<
-  WithoutEmotionSpecific<MuiCheckboxProps>,
-  'size' | 'color'
-> & {
-  /**
-   * Флаг для активации error стилей.
-   */
-  isError?: boolean;
-};
