+++ packages/components/src/Chevron/Chevron.tsx
@@ -1,5 +1,13 @@
-import { ChevronProps } from './types';
+import { SvgIconProps } from '@mui/material';
+
+import { WithoutEmotionSpecific } from '../types';
+
 import { ChevronWrapper } from './styles';
+
+export type ChevronProps = { isActive?: boolean; className?: string } & Omit<
+  WithoutEmotionSpecific<SvgIconProps>,
+  'children' | 'classes' | 'ref'
+>;
 
 export const Chevron = (props: ChevronProps): JSX.Element => (
   <ChevronWrapper {...props} />
