+++ packages/components/src/Chevron/index.ts
@@ -1,3 +1,1 @@
 export * from './Chevron';
-
-export * from './types';
