--- packages/components/src/Chevron/types.ts
@@ -1,8 +1,0 @@
-import { SvgIconProps } from '@mui/material';
-
-import { WithoutEmotionSpecific } from '../types';
-
-export type ChevronProps = { isActive?: boolean; className?: string } & Omit<
-  WithoutEmotionSpecific<SvgIconProps>,
-  'children' | 'classes' | 'ref'
->;
