+++ packages/components/src/DatePicker/DateCalendar/DateCalendarGridBtn/DateCalendarGridBtn.tsx
@@ -1,6 +1,3 @@
 import { DateCalendarGridBtnWrapper } from './styles';
-import { DateCalendarDayBtnWrapperProps } from './types';
 
-export const DateCalendarGridBtn = (props: DateCalendarDayBtnWrapperProps) => (
-  <DateCalendarGridBtnWrapper {...props} />
-);
+export const DateCalendarGridBtn = DateCalendarGridBtnWrapper;
