+++ packages/components/src/DatePicker/DateCalendar/DateCalendarGridBtn/styles.ts
@@ -1,8 +1,35 @@
+import { ReactNode } from 'react';
+
 import { styled } from '../../../styles';
 import { DateCalendarButton } from '../DateCalendarBtn';
 import { Theme } from '../../../theme';
+import { ButtonProps } from '../../../Button';
 
-import { DateCalendarDayBtnWrapperProps } from './types';
+export type DateCalendarDayBtnWrapperProps = Omit<ButtonProps, 'variant'> & {
+  /**
+   * @description флаг означающий, что дата находится вне доступного диапазона,
+   * Если принимает значение true, то это значит, что дата находится вне целевого диапазона, предназначенного для выбора. Текст кнопки будет иметь серый оттенок.
+   * Если принимает значение false, то это значит, что дата находится внутри целевого диапазона, предназначенного для выбора. Текст кнопки будет иметь стандартный оттенок.
+   */
+  isOutOfAvailableRange?: boolean;
+  /**
+   * @description флаг означающий, что дата совпадает с локальным временем пользователя
+   */
+  isCurrentInUserLocalTime: boolean;
+  /**
+   * @description флаг означающий, что дата находится в выбранном диапазоне
+   */
+  isInSelectedRange?: boolean;
+  /**
+   * @description количество элементов в строке
+   */
+  lengthInRow: number;
+  children: ReactNode;
+  /**
+   * @description флаг, означающий, что предыдущий элемент находится в выбранном диапазоне
+   */
+  isPreviousItemInSelectedRange?: boolean;
+};
 
 type WithTheme = {
   theme: Theme;
