+++ packages/components/src/Fab/Fab.tsx
@@ -1,8 +1,19 @@
+import { FabProps as MuiFabProps } from '@mui/material/Fab/Fab';
 import { forwardRef } from 'react';
 
+import { WithoutEmotionSpecific } from '../types';
+
 import { StyledFab } from './styles';
-import { FabProps } from './types';
+import { FabColor } from './types';
 import { FabSizes } from './enums';
+
+export type FabProps = Omit<
+  WithoutEmotionSpecific<MuiFabProps>,
+  'color' | 'variant'
+> & {
+  color?: FabColor;
+  variant?: 'circular' | 'square';
+};
 
 export const Fab = forwardRef<HTMLButtonElement, FabProps>(
   (
