+++ packages/components/src/Grid/Grid.tsx
@@ -1,3 +1,6 @@
-import { StyledGrid } from './styles';
+import { StyledGrid, StyledGridProps } from './styles';
+import { GridComponentProps } from './GridComponent';
+
+export type GridProps = GridComponentProps & StyledGridProps;
 
 export const Grid = StyledGrid;
