+++ packages/components/src/Grid/GridComponent.tsx
@@ -1,8 +1,24 @@
-import { forwardRef } from 'react';
+import {
+  DetailedHTMLProps,
+  ElementType,
+  HTMLAttributes,
+  ReactNode,
+  forwardRef,
+} from 'react';
 
-import { GridProps } from './types';
+export type GridComponentProps = DetailedHTMLProps<
+  HTMLAttributes<HTMLDivElement>,
+  HTMLDivElement
+> & {
+  /**
+   * Тип элемента
+   * @default 	'div'
+   */
+  component?: ElementType;
+  children?: ReactNode;
+};
 
-export const GridComponent = forwardRef<HTMLDivElement, GridProps>(
+export const GridComponent = forwardRef<HTMLDivElement, GridComponentProps>(
   ({ component: Component = 'div', ...props }, ref) => {
     return <Component ref={ref} {...props} />;
   },
