+++ packages/components/src/Grid/index.ts
@@ -1,3 +1,1 @@
 export * from './Grid';
-
-export * from './types';
