+++ packages/components/src/OverflowTypography/OverflowTypography.tsx
@@ -1,10 +1,38 @@
 import { forwardRef } from 'react';
+import { PropsWithChildren } from 'react';
 
 import { Tooltip } from '../Tooltip';
+import { TooltipProps as BasicTooltipProps } from '../Tooltip';
+import { TypographyProps } from '../Typography';
 
 import { OverflowTypographyWrapper } from './styles';
 import { useOverflowed } from './hooks/useOverflowed';
-import { OverflowedTypographyProps } from './types';
+
+type TooltipProps = Omit<BasicTooltipProps, 'ref'>;
+
+export type OverflowedProps = {
+  /**
+   * @example <OverflowTypography rowsCount={2} />
+   * @default 1
+   * @description опорная единица по которой определяется максимиально отображаемое колличество строк
+   */
+  rowsCount?: number;
+};
+
+type TooltipCustomizable = {
+  /**
+   * @example <OverflowTypography tooltipProps={{placement: 'top-start'}} />
+   * @description способ кастомизировать тултип при необходимости
+   */
+  tooltipProps?: Omit<TooltipProps, 'children'>;
+};
+
+export type OverflowedElementProps = OverflowedProps &
+  TooltipCustomizable &
+  TypographyProps;
+
+export type OverflowedTypographyProps =
+  PropsWithChildren<OverflowedElementProps>;
 
 export const DEFAULT_ROWS_COUNT = 1;
