+++ packages/components/src/OverflowTypography/index.ts
@@ -1,3 +1,1 @@
 export * from './OverflowTypography';
-
-export * from './types';
