+++ packages/components/src/Radio/Radio.tsx
@@ -1,8 +1,20 @@
 import { forwardRef } from 'react';
+import { RadioProps as MuiRadioProps } from '@mui/material/Radio/Radio';
+
+import { WithoutEmotionSpecific } from '../types';
 
 import { StyledRadio } from './styles';
 import { CheckedCircleIcon, DefaultIcon } from './Icon';
-import { RadioProps } from './types';
+
+export type RadioProps = Omit<
+  WithoutEmotionSpecific<MuiRadioProps>,
+  'size' | 'color'
+> & {
+  /**
+   * Флаг для применения error стилей.
+   */
+  isError?: boolean;
+};
 
 export const Radio = forwardRef<HTMLButtonElement, RadioProps>(
   (
@@ -18,5 +30,3 @@ export const Radio = forwardRef<HTMLButtonElement, RadioProps>(
     );
   },
 );
-
-export default Radio;
