+++ packages/components/src/Radio/index.ts
@@ -1,3 +1,1 @@
 export * from './Radio';
-
-export * from './types';
