+++ packages/components/src/Radio/styles.ts
@@ -1,10 +1,18 @@
 import { Radio, css } from '@mui/material';
+import { RadioProps as MuiRadioProps } from '@mui/material/Radio/Radio';
 
 import { styled } from '../styles';
 import { Theme } from '../theme';
+import { WithoutEmotionSpecific } from '../types';
 
-import { RadioProps } from './types';
 import { CHECKED_ICON_CLASSES, DEFAULT_ICON_CLASSES } from './constants';
+
+type StyledRadioProps = Omit<
+  WithoutEmotionSpecific<MuiRadioProps>,
+  'size' | 'color'
+> & {
+  isError?: boolean;
+};
 
 const getErrorStyles = (theme: Theme) => css`
   &.MuiRadio-root:not(.Mui-disabled) {
@@ -20,7 +28,7 @@ const getErrorStyles = (theme: Theme) => css`
 
 export const StyledRadio = styled(Radio, {
   shouldForwardProp: (prop) => prop !== 'isError',
-})<RadioProps>`
+})<StyledRadioProps>`
   padding: 5px;
 
   color: ${({ theme }) => theme.palette.grey[300]};
