+++ packages/components/src/Tag/Tag.tsx
@@ -1,10 +1,51 @@
-import { forwardRef } from 'react';
+import { ReactElement, forwardRef } from 'react';
 import { CrossSmOutlineSm } from '@astral/icons';
+import { ChipProps as MuiTagProps } from '@mui/material/Chip/Chip';
 
 import { LegacyGrid } from '../LegacyGrid';
+import { WithoutEmotionSpecific } from '../types';
+import { BadgeColor } from '../Badge';
 
 import { StyledTag, getBadgeColor } from './styles';
-import { TagProps } from './types';
+import { TagColor, TagSize, TagVariant } from './types';
+
+export type TagAddonProps = {
+  color?: BadgeColor;
+};
+
+export type TagAddon = (props: TagAddonProps) => ReactElement | null;
+
+export type TagProps = Omit<
+  WithoutEmotionSpecific<MuiTagProps>,
+  'color' | 'variant' | 'size'
+> & {
+  /**
+   * Цвет тега
+   */
+  color?: TagColor;
+  /**
+   * Тип тега
+   */
+  variant?: TagVariant;
+  /**
+   * Скругленная форма тега
+   */
+  rounded?: boolean;
+  /**
+   * Размер тега
+   */
+  size?: TagSize;
+
+  /**
+   * Контент слева от label
+   */
+  startAddon?: TagAddon;
+
+  /**
+   * Контент справа от label
+   */
+  endAddon?: TagAddon;
+};
 
 export const Tag = forwardRef<HTMLDivElement, TagProps>(
   (
